// apps/web/src/proto/Client.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Platform } from "./Platform";

export type Client = { id: number, uid: string, nickname: string, 
/**
 * Voice channel; `None` while on the server without being in voice.
 */
channel: number | null, groups: Array<number>, platform: Platform, muted: boolean, deafened: boolean, away: string | null, talking: boolean, 
/**
 * Whether the client currently has a connected audio transport.
 */
voice: boolean, };